Improve Logging and code readability
The static method `ChairManager#getLogger()` ensures that a valid Logger can be used without null-checking and while making sure that the plugin's logger is used if possible.

Of course this method should not be used by third-party plugins but if they wanted to, they could just use `JavaPlugin#getLogger()` anyway.