Adds setting to disable permission check for regeneration effect (#108)
The default value is set to `true` for now so existing users don't get the behaviour changed.
If I don't forget it, I'll change this to `false` for the next release as I think it is a more desired default value. But regeneration is disabled by default too, so they should see the setting and configure it to their likeing :shrug: