Fixes players falling through blocks when no tp to old location (#112)
When `TeleportPlayerToOldLocation` is turned off
in the configuration file, players are falling through the block
into the ground (or into a cave etc.).

The player is now always being teleported (if not asked not to via API call)
but the `Chair#getPlayerLeavingLocation()`
takes that setting into account and produces a
fitting Location.

Tested on Spigot 1.19.4 (3684-Spigot-41150f3-f770708)